Add blake3 hash to meta-data during pack
When packing a file or block device, also calculate its blake3 hash.
The hash is stored in config.yaml for each stream. This allows us to
verify during unpack that the data written matches what was packed.

The verify command gains a new '--internal' option which compares the
stored hash with the hash computed from the stream bytes.

Example:
  duplocalypse archive verify [OPTIONS] --archive <ARCHIVE> --stream <STREAM> \
    <INPUT|--internal>

With this option, the original file or device is no longer required to
validate the expanded stream integrity.

Signed-off-by: Tony Asleson <tasleson@redhat.com>