[cli] Add ability to pass multiple input names
Add the ability to have multiple input files on the command
line.  Doing this we can reduce the creation of small slabs
by keeping the archive and chunker open across files.

Note: This isn't ideal yet as when we're using compression
the slab size on disk can be smaller than target size of
4MiB so the slab sizes end up being whatever the compression
ratio we are getting.

Signed-off-by: Tony Asleson <tasleson@redhat.com>